% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_tips}
\alias{find_tips}
\title{Find tip taxon names matching a pattern}
\usage{
find_tips(geo, pattern, fixed = FALSE)
}
\arguments{
\item{geo}{A geo_ancstates object.}

\item{pattern}{Character. Regex or fixed string.}

\item{fixed}{Logical. Fixed string match? (default FALSE).}
}
\description{
Find tip taxon names matching a pattern
}
