% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_node_ages}
\alias{get_node_ages}
\title{List all unique node ages (rounded to integer Ma)}
\usage{
get_node_ages(geo)
}
\arguments{
\item{geo}{A geo_ancstates object.}
}
\description{
List all unique node ages (rounded to integer Ma)
}
